Allow entity ids to be strings in datoms
We are using wanderung for rewriting history (i.e. adding new datoms with tempids), and one problem with existing spec is that is disallows tempids, while the actual migration process seem to work fine.

I'm not sure whether support for rewriting history is in scope for wanderung, if it isn't, feel free to decline the PR!